backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0

# Database
mysqlclient==2.2.4
dj-database-url==2.2.0
psycopg2-binary==2.9.9

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0

# Storage & Media
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.2

# API Documentation
drf-yasg==1.21.7
django-filter==23.5

# Utilities
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
